refactor(gui): stage 3 — extract sidebar module
src/app/sidebar.js: renderSidebar, updateSidebarSelection,
renderProject/renderSession, drag-reorder math, inline renames — moved
verbatim. View/focus callbacks (switchTo, switchToProject,
confirmAndDeleteProject, renderEmptyState, refocusActiveTerm) injected
via initSidebar(deps); modal opens imported directly (sidebar → modals
is acyclic).

Suites: vitest 149, mock e2e 52(+1 skip), e2e-real 5.
